+++ babel.config.js
@@ -1,5 +1,18 @@
 const getPresets = isDev => [
-  '@babel/preset-env',
+  [
+    '@babel/preset-env',
+    {
+      targets: {
+        ie: '11',
+        edge: '17',
+        firefox: '60',
+        chrome: '67',
+        safari: '11.1',
+      },
+      useBuiltIns: 'usage',
+      corejs: '3.6.4',
+    },
+  ],
   '@babel/preset-react',
   '@babel/preset-typescript',
   [