+++ examples/create-react-app/src/index.tsx
@@ -1,7 +1,11 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import {Theme} from '@twilio-paste/theme';
-import {Box, Heading, Button} from '@twilio-paste/core';
+import {Theme} from '@twilio-paste/core/theme';
+import {Anchor} from '@twilio-paste/core';
+import {Box} from '@twilio-paste/core/box';
+import {Button} from '@twilio-paste/core/button';
+import {Heading} from '@twilio-paste/core/heading';
+import {PlusIcon} from '@twilio-paste/icons/esm/PlusIcon';
 import * as serviceWorker from './serviceWorker';
 
 function App() {
@@ -9,9 +13,12 @@ function App() {
     <Theme.Provider theme="default">
       <Box margin="space60">
         <Heading as="h1" variant="heading10">
-          Start here.
+          <Anchor href="https://paste.twilio.design">Start here.</Anchor>
         </Heading>
-        <Button variant="primary">Primary button</Button>
+        <Button variant="primary">
+          <PlusIcon decorative />
+          Primary button
+        </Button>
       </Box>
     </Theme.Provider>
   );