+++ packages/paste-core/components/disclosure/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.31](https://github.com/twilio-labs/paste/compare/@twilio-paste/disclosure@0.2.30...@twilio-paste/disclosure@0.2.31) (2020-10-13)
+
+**Note:** Version bump only for package @twilio-paste/disclosure
+
+
+
+
+
 ## [0.2.30](https://github.com/twilio-labs/paste/compare/@twilio-paste/disclosure@0.2.29...@twilio-paste/disclosure@0.2.30) (2020-10-09)
 
 **Note:** Version bump only for package @twilio-paste/disclosure