+++ packages/paste-core/components/heading/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.46](https://github.com/twilio-labs/paste/compare/@twilio-paste/heading@2.0.45...@twilio-paste/heading@2.0.46) (2020-10-13)
+
+**Note:** Version bump only for package @twilio-paste/heading
+
+
+
+
+
 ## [2.0.45](https://github.com/twilio-labs/paste/compare/@twilio-paste/heading@2.0.44...@twilio-paste/heading@2.0.45) (2020-10-07)
 
 **Note:** Version bump only for package @twilio-paste/heading