+++ packages/paste-core/components/input/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.4](https://github.com/twilio-labs/paste/compare/@twilio-paste/input@1.1.3...@twilio-paste/input@1.1.4) (2020-10-13)
+
+**Note:** Version bump only for package @twilio-paste/input
+
+
+
+
+
 ## [1.1.3](https://github.com/twilio-labs/paste/compare/@twilio-paste/input@1.1.2...@twilio-paste/input@1.1.3) (2020-10-07)
 
 **Note:** Version bump only for package @twilio-paste/input