+++ packages/paste-core/components/screen-reader-only/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.13](https://github.com/twilio-labs/paste/compare/@twilio-paste/screen-reader-only@1.1.12...@twilio-paste/screen-reader-only@1.1.13) (2020-10-13)
+
+**Note:** Version bump only for package @twilio-paste/screen-reader-only
+
+
+
+
+
 ## [1.1.12](https://github.com/twilio-labs/paste/compare/@twilio-paste/screen-reader-only@1.1.11...@twilio-paste/screen-reader-only@1.1.12) (2020-08-19)
 
 **Note:** Version bump only for package @twilio-paste/screen-reader-only