+++ packages/paste-core/components/separator/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.14](https://github.com/twilio-labs/paste/compare/@twilio-paste/separator@0.2.13...@twilio-paste/separator@0.2.14) (2020-10-13)
+
+**Note:** Version bump only for package @twilio-paste/separator
+
+
+
+
+
 ## [0.2.13](https://github.com/twilio-labs/paste/compare/@twilio-paste/separator@0.2.12...@twilio-paste/separator@0.2.13) (2020-10-07)
 
 **Note:** Version bump only for package @twilio-paste/separator