+++ packages/paste-core/layout/media-object/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.65](https://github.com/twilio-labs/paste/compare/@twilio-paste/media-object@1.2.64...@twilio-paste/media-object@1.2.65) (2020-10-13)
+
+**Note:** Version bump only for package @twilio-paste/media-object
+
+
+
+
+
 ## [1.2.64](https://github.com/twilio-labs/paste/compare/@twilio-paste/media-object@1.2.63...@twilio-paste/media-object@1.2.64) (2020-10-07)
 
 **Note:** Version bump only for package @twilio-paste/media-object