+++ packages/paste-icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.6.1](https://github.com/twilio-labs/paste/compare/@twilio-paste/icons@3.6.0...@twilio-paste/icons@3.6.1) (2020-10-13)
+
+**Note:** Version bump only for package @twilio-paste/icons
+
+
+
+
+
 # [3.6.0](https://github.com/twilio-labs/paste/compare/@twilio-paste/icons@3.5.8...@twilio-paste/icons@3.6.0) (2020-10-09)